requirements.txt
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-auth-httplib2==0.3.0
google-genai==1.55.0
moviepy==2.2.1
openai-whisper==20250625
edge-tts==7.2.7
groq==1.0.0
python-dotenv==1.2.1
requests==2.32.5
openai==2.15.0